fixed some compile issues in tests (didn't check if tests passed, just compiled)

+++ app/src/main/java/com/seniordesign/wolfpack/quizinator/Database/Rules/Rules.java
@@ -69,5 +69,5 @@ public class Rules {
 
     public long getDeckId() { return deckId; }
 
-    public void setDeckId(int deckId) { this.deckId = deckId; }
+    public void setDeckId(long deckId) { this.deckId = deckId; }
 }
